+++ examples/basic-vite/src/Layout/index.tsx
@@ -1,0 +1,8 @@
+export default ({ children }) => {
+  return (
+    <div>
+      Layout
+      {children}
+    </div>
+  );
+};