+++ packages/build-mpa-config/src/index.ts
@@ -38,10 +38,12 @@ export const generateMPAEntries = (api: IPluginAPI, options: IConfigOptions) => 
     // when the entry has no export default declaration, do not generate any files
     let finalEntry = entryPath;
     let runAppPath = null;
+    let routesFilePath;
     if (isAppEntry || checkExportDefaultDeclarationExists(path.join(rootDir, 'src', source))) {
       const result = generateEntry(api, { framework, targetDir, pageEntry: entryPath, entryName, pageConfig, isAppEntry });
       finalEntry = result.entryPath;
       runAppPath = result.runAppPath;
+      routesFilePath = result.routesFilePath;
     }
 
     parsedEntries[entryName] = {
@@ -49,6 +51,7 @@ export const generateMPAEntries = (api: IPluginAPI, options: IConfigOptions) => 
       finalEntry,
       shouldRedirectRunApp: isAppEntry,
       runAppPath,
+      routesFilePath,
     };
   });
   return parsedEntries;
@@ -67,17 +70,17 @@ const setMPAConfig = (api, config, options: IConfigOptions) => {
   // add redirect entry path
   const redirectEntries: IGenerateResult[] = [];
   Object.keys(parsedEntries).forEach((entryKey) => {
-    const { entryName, source, finalEntry, shouldRedirectRunApp, runAppPath } = parsedEntries[entryKey];
+    const { entryName, finalEntry, shouldRedirectRunApp, runAppPath, routesFilePath } = parsedEntries[entryKey];
     config.entry(entryName).add(finalEntry);
     if (shouldRedirectRunApp) {
       redirectEntries.push({
         entryPath: finalEntry,
         runAppPath,
+        routesFilePath,
       });
     }
     // get page paths for rule match
-    const matchStr = `src/${source}`;
-    matchStrs.push(formatPath(matchStr));
+    matchStrs.push(formatPath(routesFilePath));
   });
 
   api.applyMethod('addImportDeclaration', {