+++ packages/build-mpa-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.0.2
+
+- fix match rules for app.json
+
 ## 4.0.1
 
 - feat add renderData of `isMPA`