+++ packages/plugin-moment-locales/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.3
+
+- [fix] pre build moment locales
+
+## 0.1.2
+
+- [fix] generate sourceMap if necessary
+
 ## 0.1.1
 
 - [feat] support inject moment locales in vite mode